Remove unnecessary repository factory interface and move it to domain
