+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2025-05-16T08:28:34.987Z\n"
-"PO-Revision-Date: 2025-05-16T08:28:34.987Z\n"
+"POT-Creation-Date: 2025-05-16T08:40:34.424Z\n"
+"PO-Revision-Date: 2025-05-16T08:40:34.424Z\n"
 
 msgid ""
 "THIS NEW RELEASE INCLUDES SHARING SETTINGS PER INSTANCES. FOR THIS VERSION "
