+++ src/data/metadata/__tests__/integration/sync-metadata.spec.ts
@@ -8,7 +8,6 @@ import { MetadataSyncUseCase } from "../../../../domain/metadata/usecases/Metada
 import { SynchronizationBuilder } from "../../../../domain/synchronization/entities/SynchronizationBuilder";
 import { registerDynamicRepositoriesInFactory } from "../../../../presentation/CompositionRoot";
 import { startDhis } from "../../../../utils/dhisServer";
-import { DefaultDynamicRepositoryFactory } from "../../../common/factories/DefaultDynamicRepositoryFactory";
 
 const repositoryFactory = buildRepositoryFactory();
 
@@ -216,7 +215,7 @@ describe("Sync metadata", () => {
 });
 
 function buildRepositoryFactory() {
-    const repositoryFactory: DynamicRepositoryFactory = new DefaultDynamicRepositoryFactory();
+    const repositoryFactory: DynamicRepositoryFactory = new DynamicRepositoryFactory();
 
     registerDynamicRepositoriesInFactory(repositoryFactory);
